[PR feedback] Split out new `_breadcrumb_content` file + move shared types into a separate file
- the breadcrumb types weren't super married to the actual component usage in any case

+ move unit tests for content to its own file, add new basic it renders test for other components
